1. В UserEntity сохраняется путь к аватарке, а не ее id. 2. В UserDTO удалил поле photo, мапперы переделал под поле image. 3. Изменил методы для сохранения аватарок пользователя. 4. Добавил changeset для изменения таблицы users 5. В application.properties необходимо прописать путь к папке с аватарками

+++ src/main/java/ru/skypro/homework/service/UserService.java
@@ -14,6 +14,6 @@ public interface UserService {
     UserEntity getAuthorizedUser(Authentication authentication);
     void changePassword(NewPasswordDTO newPasswordDTO, Authentication authentication);
     void updateAvatar(MultipartFile image, Authentication authentication) throws IOException;
-    byte[] getUserImage(Long userId) throws IOException;
+    byte[] getUserImage(Integer userId) throws IOException;
 
 }
